+++ README.md
@@ -1,0 +1,15 @@
+# PoGo Rarity
+
+Utilities for aggregating Pokémon GO rarity information.
+
+## Usage
+
+Install the package and run the CLI:
+
+```bash
+pokemon-rarity --limit 50
+```
+
+The `--limit` option restricts the number of Pokémon processed from
+PokemonDB, which is useful during testing.
+
